Fixed Problems with Sidenav Loading
The sidenav had problems rendering properly because it relies on the entire document
being rendered to know where the header and the footer are. Chrome worked fine, but
Firefox did not. Now, we're loading the scripts at the end of the body and we aren't
using async, which we'll need to fix later.